+++ core/src/main/java/ru/ra_tech/garden_manager/core/controllers/error_responses/EntityNotFoundResponse.java
@@ -6,7 +6,7 @@ import org.springframework.http.ProblemDetail;
 public class EntityNotFoundResponse extends AbstractErrorResponse {
     private static final String MESSAGE = "%s with id #%d was not found";
 
-    public EntityNotFoundResponse(String entity, int id) {
+    public EntityNotFoundResponse(String entity, long id) {
         super(ProblemDetail.forStatusAndDetail(
                 HttpStatus.NOT_FOUND,
                 String.format(MESSAGE, entity, id)
