+++ core/src/main/java/ru/ra_tech/garden_manager/core/security/JwtPrincipal.java
@@ -1,4 +1,4 @@
 package ru.ra_tech.garden_manager.core.security;
 
-public record JwtPrincipal(int id, String login, String name) {
+public record JwtPrincipal(long id, String login, String name) {
 }
