+++ database/src/main/java/ru/ra_tech/garden_manager/database/repositories/garden/CreateGardenDto.java
@@ -2,5 +2,5 @@ package ru.ra_tech.garden_manager.database.repositories.garden;
 
 import org.springframework.lang.Nullable;
 
-public record CreateGardenDto(String name, @Nullable String address, int ownerId) {
+public record CreateGardenDto(String name, @Nullable String address, long ownerId) {
 }
