+++ database/src/main/java/ru/ra_tech/garden_manager/database/repositories/garden/GardenUsersDto.java
@@ -2,4 +2,4 @@ package ru.ra_tech.garden_manager.database.repositories.garden;
 
 import io.vavr.collection.List;
 
-public record GardenUsersDto(int id, int ownerId, List<Integer> participants) {}
+public record GardenUsersDto(long id, long ownerId, List<Long> participants) {}
