+++ database/src/main/java/ru/ra_tech/garden_manager/database/repositories/user/UserDto.java
@@ -2,5 +2,5 @@ package ru.ra_tech.garden_manager.database.repositories.user;
 
 import org.springframework.lang.Nullable;
 
-public record UserDto(int id, String login, String name, @Nullable String email) {
+public record UserDto(long id, String login, String name, @Nullable String email) {
 }
